requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
scikit-learn==1.4.2
pandas
numpy
python-multipart
pymysql
python-dotenv
google